LU-19478 nodemap: skip config distribution for ro commands
"lctl nodemap_test_nid" and "lctl nodemap_test_id" are nodemap
commands which do not modify the nodemap config and are therefore
read-only. However, if run from the MGS, "mgs_revoke_lock()" is called
which notifies all targets to fetch the current config from the MGS,
called nodemap synchronization.

When only persistent nodemaps are in use, this is an unnecessary
communication step. But, when dynamic nodemaps are set on MDSs or
OSSs, this results in wiping all dynamic nodemaps on the targets since
the MGS's nodemap config overwrites the in-memory nodemap on the
targets.

This patch skips nodemap synchronization for read-only nodemap
operations. Note, this patch does not skip synchronization for nodemap
commands that could update the nodemap config. For instance, two
consecutive "lctl nodemap_active 1" commands (within separate nodemap
synchronization steps) still results in nodemap synchronization.

Test-Parameters: trivial testlist=sanity-sec
Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: I3af4f75056f23ef7f7432a096a7eb98ec48ae327